Split core definitions into modules
