+++ game/constants.py
@@ -1,0 +1,83 @@
+# ============================================================================
+# 高解析度設定 - 支援1920x1080及以上
+# ============================================================================
+
+# 視窗設定 - 預設Full HD
+WINDOW_WIDTH = 1920
+WINDOW_HEIGHT = 1080
+FPS = 60
+
+# 支援的解析度
+RESOLUTIONS = [
+    (1280, 720),   # HD
+    (1600, 900),   # HD+
+    (1920, 1080),  # Full HD
+    (2560, 1440),  # 2K
+    (3840, 2160),  # 4K
+]
+
+# 遊戲網格設定
+TILE_SIZE = 48  # 提升瓦片大小以適應高解析度
+MAP_WIDTH = 60
+MAP_HEIGHT = 45
+
+# 小地圖設定
+MINIMAP_SIZE = 300
+MINIMAP_TILE_SIZE = 5
+
+# 訊息日誌設定
+MAX_LOG_MESSAGES = 25
+MESSAGE_DISPLAY_TIME = 12000
+
+# ============================================================================
+# 現代化配色方案 - 高對比度設計
+# ============================================================================
+
+class Colors:
+    # 基礎色彩
+    BLACK = (0, 0, 0)
+    WHITE = (255, 255, 255)
+    
+    # 主題色彩 - 霓虹風格
+    NEON_BLUE = (0, 255, 255)
+    NEON_PINK = (255, 0, 128)
+    NEON_GREEN = (0, 255, 128)
+    NEON_YELLOW = (255, 255, 0)
+    NEON_PURPLE = (255, 0, 255)
+    NEON_ORANGE = (255, 128, 0)
+    NEON_CYAN = (0, 255, 255)
+    
+    # 遊戲色彩
+    HEALTH_RED = (255, 60, 90)
+    MANA_BLUE = (60, 180, 255)
+    EXP_YELLOW = (255, 215, 0)
+    POISON_GREEN = (128, 255, 0)
+    
+    # 地形色彩 - 高對比度
+    FLOOR_COLOR = (40, 40, 50)
+    WALL_COLOR = (20, 20, 25)
+    FLOOR_HIGHLIGHT = (60, 60, 75)
+    WALL_HIGHLIGHT = (35, 35, 40)
+    STAIRS_COLOR = (255, 200, 50)
+    
+    # UI色彩
+    UI_BG = (15, 15, 20)
+    UI_SECONDARY = (25, 25, 35)
+    UI_BORDER = (100, 100, 120)
+    UI_HIGHLIGHT = (0, 200, 255)
+    UI_TEXT = (220, 220, 230)
+    UI_TEXT_DARK = (150, 150, 160)
+    
+    # 特效色彩
+    PARTICLE_FIRE = [(255, 100, 0), (255, 150, 0), (255, 200, 0), (255, 255, 100)]
+    PARTICLE_ICE = [(100, 200, 255), (150, 220, 255), (200, 240, 255), (255, 255, 255)]
+    PARTICLE_HEAL = [(0, 255, 100), (100, 255, 150), (200, 255, 200)]
+    
+    # 訊息色彩
+    MSG_NORMAL = (200, 200, 210)
+    MSG_COMBAT = (255, 100, 100)
+    MSG_ITEM = (100, 255, 150)
+    MSG_LEVEL = (255, 220, 100)
+    MSG_STORY = (150, 200, 255)
+    MSG_WARNING = (255, 150, 50)
+
